requirements.txt
asgiref==3.2.10
autopep8==1.5.5
certifi==2020.12.5
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
django-tinymce==3.2.0
gunicorn==20.0.4
idna==2.8
Pillow==8.1.0
psycopg2==2.7.7
pycodestyle==2.6.0
pytz==2020.5
requests==2.22.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.11
wayforpay==1.0.1
whitenoise==5.2.0
